+++ app/proposal.go
@@ -276,6 +276,9 @@ func (h *ProposalHandler) SetBlockList(blob []byte) error {
 		if err != nil {
 			return fmt.Errorf("invalid bech32 address: %s, err: %w", s, err)
 		}
+		if IsUnblockable(addr) {
+			continue
+		}
 		encoded, err := h.addressCodec.BytesToString(addr)
 		if err != nil {
 			return fmt.Errorf("invalid bech32 address: %s, err: %w", s, err)
@@ -339,8 +342,12 @@ func (h *ProposalHandler) ValidateTransaction(tx sdk.Tx, txBz []byte) error {
 				for _, auth := range ethTx.SetCodeAuthorizations() {
 					addr, err := auth.Authority()
 					if err == nil {
-						if _, ok := h.blocklist[sdk.AccAddress(addr.Bytes()).String()]; ok {
-							return fmt.Errorf("signer is blocked: %s", addr.String())
+						encoded, err := h.addressCodec.BytesToString(addr.Bytes())
+						if err != nil {
+							return fmt.Errorf("invalid bech32 address: %s, err: %w", addr, err)
+						}
+						if _, ok := h.blocklist[encoded]; ok {
+							return fmt.Errorf("signer is blocked: %s", encoded)
 						}
 					}
 					// check the target address